Camera Class

+++ src/VertexArray.cpp
@@ -27,7 +27,7 @@ void VertexArray::AddBuffer(const VertexBuffer& vb, const VertexBufferLayout& la
 	vb.Bind();
 	const auto& elements = layout.GetElements();
 	unsigned int offset = 0;
-	for (size_t i = 0; i < elements.size(); i++)
+	for (unsigned int i = 0; i < elements.size(); i++)
 	{
 		const auto element = elements[i];
 		GLCall(glVertexAttribPointer(i, element.count, element.type,
